% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasEvaluation.R
\name{anzahlIndikatorenProEffekt}
\alias{anzahlIndikatorenProEffekt}
\title{Anzahl an Indikatoren pro Effekt}
\usage{
anzahlIndikatorenProEffekt(
  indikatoren = kuras_indikator(),
  effekte = kuras_effekt()
)
}
\arguments{
\item{indikatoren}{data frame as returned by \code{\link{kuras_indikator}}}

\item{effekte}{data frame as returned by \code{\link{kuras_effekt}}}
}
\value{
data frame mit Spalten \emph{eID} (Effekt-ID), \emph{eName}
  (Effekt-Name), \emph{x} (Anzahl Indikatoren fuer den Effekt).
}
\description{
Anzahl an Indikatoren pro Effekt
}
