% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasEvaluation.R
\name{beispielAuswertung.1}
\alias{beispielAuswertung.1}
\title{Beispiel-Auswertung 1}
\usage{
beispielAuswertung.1()
}
\description{
Beispielauswertung 1: Anzahl Indikatoren je Effekt und Anzahl Literaturwerte
  je Indikator
}
