% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_anlageeigenschaft}
\alias{kuras_anlageeigenschaft}
\title{select records from tblanlageeigenschaft}
\usage{
kuras_anlageeigenschaft(
  where = NULL,
  where_aeID = NA,
  where_aeName = NA,
  where_aeEinheit = NA,
  where_aeBeschreibung = NA,
  select = NULL,
  select_aeID = NA,
  select_aeName = NA,
  select_aeEinheit = NA,
  select_aeBeschreibung = NA,
  orderBy = NULL,
  orderBy_aeID = NA,
  orderBy_aeName = NA,
  orderBy_aeEinheit = NA,
  orderBy_aeBeschreibung = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_aeID}{condition for field \emph{aeID}: Enter a field description here}

\item{where_aeName}{condition for field \emph{aeName}: Enter a field description here}

\item{where_aeEinheit}{condition for field \emph{aeEinheit}: Enter a field description here}

\item{where_aeBeschreibung}{condition for field \emph{aeBeschreibung}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_aeID}{\code{select} field \emph{aeID}?}

\item{select_aeName}{\code{select} field \emph{aeName}?}

\item{select_aeEinheit}{\code{select} field \emph{aeEinheit}?}

\item{select_aeBeschreibung}{\code{select} field \emph{aeBeschreibung}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_aeID}{order by field \emph{aeID}?}

\item{orderBy_aeName}{order by field \emph{aeName}?}

\item{orderBy_aeEinheit}{order by field \emph{aeEinheit}?}

\item{orderBy_aeBeschreibung}{order by field \emph{aeBeschreibung}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblanlageeigenschaft
}
