% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_anlageeigenschaftswert}
\alias{kuras_anlageeigenschaftswert}
\title{select records from tblanlageeigenschaftswert}
\usage{
kuras_anlageeigenschaftswert(
  where = NULL,
  where_awID = NA,
  where_awAnlageID = NA,
  where_awAnlageEigenschaftID = NA,
  where_awZahlwert = NA,
  where_awBeschreibung = NA,
  select = NULL,
  select_awID = NA,
  select_awAnlageID = NA,
  select_awAnlageEigenschaftID = NA,
  select_awZahlwert = NA,
  select_awBeschreibung = NA,
  orderBy = NULL,
  orderBy_awID = NA,
  orderBy_awAnlageID = NA,
  orderBy_awAnlageEigenschaftID = NA,
  orderBy_awZahlwert = NA,
  orderBy_awBeschreibung = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_awID}{condition for field \emph{awID}: Enter a field description here}

\item{where_awAnlageID}{condition for field \emph{awAnlageID}: Enter a field description here}

\item{where_awAnlageEigenschaftID}{condition for field \emph{awAnlageEigenschaftID}: Enter a field description here}

\item{where_awZahlwert}{condition for field \emph{awZahlwert}: Enter a field description here}

\item{where_awBeschreibung}{condition for field \emph{awBeschreibung}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_awID}{\code{select} field \emph{awID}?}

\item{select_awAnlageID}{\code{select} field \emph{awAnlageID}?}

\item{select_awAnlageEigenschaftID}{\code{select} field \emph{awAnlageEigenschaftID}?}

\item{select_awZahlwert}{\code{select} field \emph{awZahlwert}?}

\item{select_awBeschreibung}{\code{select} field \emph{awBeschreibung}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_awID}{order by field \emph{awID}?}

\item{orderBy_awAnlageID}{order by field \emph{awAnlageID}?}

\item{orderBy_awAnlageEigenschaftID}{order by field \emph{awAnlageEigenschaftID}?}

\item{orderBy_awZahlwert}{order by field \emph{awZahlwert}?}

\item{orderBy_awBeschreibung}{order by field \emph{awBeschreibung}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblanlageeigenschaftswert
}
