% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_anlageoderstudie}
\alias{kuras_anlageoderstudie}
\title{select records from tblanlageoderstudie}
\usage{
kuras_anlageoderstudie(
  where = NULL,
  where_aID = NA,
  where_aName = NA,
  where_aMassnahmenTypID = NA,
  where_aMassnahmenEbeneID = NA,
  where_aStandortID = NA,
  where_aAdresse = NA,
  where_aKoordinateX = NA,
  where_aKoordinateY = NA,
  where_aEntwaesserungssystem = NA,
  where_aBetreiberInstitutionID = NA,
  where_aInbetriebnahme = NA,
  where_aAusserbetriebnahme = NA,
  where_aBetriebsweise = NA,
  where_aBeschreibung = NA,
  where_aProjektID = NA,
  where_aFotoOrdner = NA,
  where_aDokumentOrdner = NA,
  where_aAngelegtAm = NA,
  where_aAngelegtVonPersonID = NA,
  where_aGeaendert = NA,
  where_aKoordinatensystem = NA,
  where_aReferenzGruppeID = NA,
  where_aIstStudie = NA,
  select = NULL,
  select_aID = NA,
  select_aName = NA,
  select_aMassnahmenTypID = NA,
  select_aMassnahmenEbeneID = NA,
  select_aStandortID = NA,
  select_aAdresse = NA,
  select_aKoordinateX = NA,
  select_aKoordinateY = NA,
  select_aEntwaesserungssystem = NA,
  select_aBetreiberInstitutionID = NA,
  select_aInbetriebnahme = NA,
  select_aAusserbetriebnahme = NA,
  select_aBetriebsweise = NA,
  select_aBeschreibung = NA,
  select_aProjektID = NA,
  select_aFotoOrdner = NA,
  select_aDokumentOrdner = NA,
  select_aAngelegtAm = NA,
  select_aAngelegtVonPersonID = NA,
  select_aGeaendert = NA,
  select_aKoordinatensystem = NA,
  select_aReferenzGruppeID = NA,
  select_aIstStudie = NA,
  orderBy = NULL,
  orderBy_aID = NA,
  orderBy_aName = NA,
  orderBy_aMassnahmenTypID = NA,
  orderBy_aMassnahmenEbeneID = NA,
  orderBy_aStandortID = NA,
  orderBy_aAdresse = NA,
  orderBy_aKoordinateX = NA,
  orderBy_aKoordinateY = NA,
  orderBy_aEntwaesserungssystem = NA,
  orderBy_aBetreiberInstitutionID = NA,
  orderBy_aInbetriebnahme = NA,
  orderBy_aAusserbetriebnahme = NA,
  orderBy_aBetriebsweise = NA,
  orderBy_aBeschreibung = NA,
  orderBy_aProjektID = NA,
  orderBy_aFotoOrdner = NA,
  orderBy_aDokumentOrdner = NA,
  orderBy_aAngelegtAm = NA,
  orderBy_aAngelegtVonPersonID = NA,
  orderBy_aGeaendert = NA,
  orderBy_aKoordinatensystem = NA,
  orderBy_aReferenzGruppeID = NA,
  orderBy_aIstStudie = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_aID}{condition for field \emph{aID}: Enter a field description here}

\item{where_aName}{condition for field \emph{aName}: Enter a field description here}

\item{where_aMassnahmenTypID}{condition for field \emph{aMassnahmenTypID}: Enter a field description here}

\item{where_aMassnahmenEbeneID}{condition for field \emph{aMassnahmenEbeneID}: Enter a field description here}

\item{where_aStandortID}{condition for field \emph{aStandortID}: Enter a field description here}

\item{where_aAdresse}{condition for field \emph{aAdresse}: Enter a field description here}

\item{where_aKoordinateX}{condition for field \emph{aKoordinateX}: Enter a field description here}

\item{where_aKoordinateY}{condition for field \emph{aKoordinateY}: Enter a field description here}

\item{where_aEntwaesserungssystem}{condition for field \emph{aEntwaesserungssystem}: Enter a field description here}

\item{where_aBetreiberInstitutionID}{condition for field \emph{aBetreiberInstitutionID}: Enter a field description here}

\item{where_aInbetriebnahme}{condition for field \emph{aInbetriebnahme}: Enter a field description here}

\item{where_aAusserbetriebnahme}{condition for field \emph{aAusserbetriebnahme}: Enter a field description here}

\item{where_aBetriebsweise}{condition for field \emph{aBetriebsweise}: Enter a field description here}

\item{where_aBeschreibung}{condition for field \emph{aBeschreibung}: Enter a field description here}

\item{where_aProjektID}{condition for field \emph{aProjektID}: Enter a field description here}

\item{where_aFotoOrdner}{condition for field \emph{aFotoOrdner}: Enter a field description here}

\item{where_aDokumentOrdner}{condition for field \emph{aDokumentOrdner}: Enter a field description here}

\item{where_aAngelegtAm}{condition for field \emph{aAngelegtAm}: Enter a field description here}

\item{where_aAngelegtVonPersonID}{condition for field \emph{aAngelegtVonPersonID}: Enter a field description here}

\item{where_aGeaendert}{condition for field \emph{aGeaendert}: Enter a field description here}

\item{where_aKoordinatensystem}{condition for field \emph{aKoordinatensystem}: Enter a field description here}

\item{where_aReferenzGruppeID}{condition for field \emph{aReferenzGruppeID}: Enter a field description here}

\item{where_aIstStudie}{condition for field \emph{aIstStudie}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_aID}{\code{select} field \emph{aID}?}

\item{select_aName}{\code{select} field \emph{aName}?}

\item{select_aMassnahmenTypID}{\code{select} field \emph{aMassnahmenTypID}?}

\item{select_aMassnahmenEbeneID}{\code{select} field \emph{aMassnahmenEbeneID}?}

\item{select_aStandortID}{\code{select} field \emph{aStandortID}?}

\item{select_aAdresse}{\code{select} field \emph{aAdresse}?}

\item{select_aKoordinateX}{\code{select} field \emph{aKoordinateX}?}

\item{select_aKoordinateY}{\code{select} field \emph{aKoordinateY}?}

\item{select_aEntwaesserungssystem}{\code{select} field \emph{aEntwaesserungssystem}?}

\item{select_aBetreiberInstitutionID}{\code{select} field \emph{aBetreiberInstitutionID}?}

\item{select_aInbetriebnahme}{\code{select} field \emph{aInbetriebnahme}?}

\item{select_aAusserbetriebnahme}{\code{select} field \emph{aAusserbetriebnahme}?}

\item{select_aBetriebsweise}{\code{select} field \emph{aBetriebsweise}?}

\item{select_aBeschreibung}{\code{select} field \emph{aBeschreibung}?}

\item{select_aProjektID}{\code{select} field \emph{aProjektID}?}

\item{select_aFotoOrdner}{\code{select} field \emph{aFotoOrdner}?}

\item{select_aDokumentOrdner}{\code{select} field \emph{aDokumentOrdner}?}

\item{select_aAngelegtAm}{\code{select} field \emph{aAngelegtAm}?}

\item{select_aAngelegtVonPersonID}{\code{select} field \emph{aAngelegtVonPersonID}?}

\item{select_aGeaendert}{\code{select} field \emph{aGeaendert}?}

\item{select_aKoordinatensystem}{\code{select} field \emph{aKoordinatensystem}?}

\item{select_aReferenzGruppeID}{\code{select} field \emph{aReferenzGruppeID}?}

\item{select_aIstStudie}{\code{select} field \emph{aIstStudie}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_aID}{order by field \emph{aID}?}

\item{orderBy_aName}{order by field \emph{aName}?}

\item{orderBy_aMassnahmenTypID}{order by field \emph{aMassnahmenTypID}?}

\item{orderBy_aMassnahmenEbeneID}{order by field \emph{aMassnahmenEbeneID}?}

\item{orderBy_aStandortID}{order by field \emph{aStandortID}?}

\item{orderBy_aAdresse}{order by field \emph{aAdresse}?}

\item{orderBy_aKoordinateX}{order by field \emph{aKoordinateX}?}

\item{orderBy_aKoordinateY}{order by field \emph{aKoordinateY}?}

\item{orderBy_aEntwaesserungssystem}{order by field \emph{aEntwaesserungssystem}?}

\item{orderBy_aBetreiberInstitutionID}{order by field \emph{aBetreiberInstitutionID}?}

\item{orderBy_aInbetriebnahme}{order by field \emph{aInbetriebnahme}?}

\item{orderBy_aAusserbetriebnahme}{order by field \emph{aAusserbetriebnahme}?}

\item{orderBy_aBetriebsweise}{order by field \emph{aBetriebsweise}?}

\item{orderBy_aBeschreibung}{order by field \emph{aBeschreibung}?}

\item{orderBy_aProjektID}{order by field \emph{aProjektID}?}

\item{orderBy_aFotoOrdner}{order by field \emph{aFotoOrdner}?}

\item{orderBy_aDokumentOrdner}{order by field \emph{aDokumentOrdner}?}

\item{orderBy_aAngelegtAm}{order by field \emph{aAngelegtAm}?}

\item{orderBy_aAngelegtVonPersonID}{order by field \emph{aAngelegtVonPersonID}?}

\item{orderBy_aGeaendert}{order by field \emph{aGeaendert}?}

\item{orderBy_aKoordinatensystem}{order by field \emph{aKoordinatensystem}?}

\item{orderBy_aReferenzGruppeID}{order by field \emph{aReferenzGruppeID}?}

\item{orderBy_aIstStudie}{order by field \emph{aIstStudie}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblanlageoderstudie
}
