% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasEvaluation.R
\name{kuras_auswertung}
\alias{kuras_auswertung}
\title{KURAS Auswertung}
\usage{
kuras_auswertung(
  effekte = kuras_effekt(stringsAsFactors = FALSE),
  indikatoren = kuras_indikator(stringsAsFactors = FALSE),
  indikatorwerte = kuras_indikatorwert(stringsAsFactors = FALSE)
)
}
\arguments{
\item{effekte}{data frame as returned by \code{\link{kuras_effekt}}}

\item{indikatoren}{data frame as returned by \code{\link{kuras_indikator}}}

\item{indikatorwerte}{data frame as returned by
\code{\link{kuras_indikatorwert}}}
}
\description{
KURAS Auswertung
}
