% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_effekt}
\alias{kuras_effekt}
\title{select records from tbleffekt}
\usage{
kuras_effekt(
  where = NULL,
  where_eID = NA,
  where_eName = NA,
  where_eBeschreibung = NA,
  select = NULL,
  select_eID = NA,
  select_eName = NA,
  select_eBeschreibung = NA,
  orderBy = NULL,
  orderBy_eID = NA,
  orderBy_eName = NA,
  orderBy_eBeschreibung = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_eID}{condition for field \emph{eID}: Enter a field description here}

\item{where_eName}{condition for field \emph{eName}: Enter a field description here}

\item{where_eBeschreibung}{condition for field \emph{eBeschreibung}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_eID}{\code{select} field \emph{eID}?}

\item{select_eName}{\code{select} field \emph{eName}?}

\item{select_eBeschreibung}{\code{select} field \emph{eBeschreibung}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_eID}{order by field \emph{eID}?}

\item{orderBy_eName}{order by field \emph{eName}?}

\item{orderBy_eBeschreibung}{order by field \emph{eBeschreibung}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tbleffekt
}
