% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_indikator}
\alias{kuras_indikator}
\title{select records from tblindikator}
\usage{
kuras_indikator(
  where = NULL,
  where_iID = NA,
  where_iEffektID = NA,
  where_iName = NA,
  where_iEinheit = NA,
  where_iBeschreibung = NA,
  where_iAngelegtAm = NA,
  where_iAngelegtVonPersonID = NA,
  where_iGeaendert = NA,
  select = NULL,
  select_iID = NA,
  select_iEffektID = NA,
  select_iName = NA,
  select_iEinheit = NA,
  select_iBeschreibung = NA,
  select_iAngelegtAm = NA,
  select_iAngelegtVonPersonID = NA,
  select_iGeaendert = NA,
  orderBy = NULL,
  orderBy_iID = NA,
  orderBy_iEffektID = NA,
  orderBy_iName = NA,
  orderBy_iEinheit = NA,
  orderBy_iBeschreibung = NA,
  orderBy_iAngelegtAm = NA,
  orderBy_iAngelegtVonPersonID = NA,
  orderBy_iGeaendert = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_iID}{condition for field \emph{iID}: Enter a field description here}

\item{where_iEffektID}{condition for field \emph{iEffektID}: Enter a field description here}

\item{where_iName}{condition for field \emph{iName}: Enter a field description here}

\item{where_iEinheit}{condition for field \emph{iEinheit}: Enter a field description here}

\item{where_iBeschreibung}{condition for field \emph{iBeschreibung}: Enter a field description here}

\item{where_iAngelegtAm}{condition for field \emph{iAngelegtAm}: Enter a field description here}

\item{where_iAngelegtVonPersonID}{condition for field \emph{iAngelegtVonPersonID}: Enter a field description here}

\item{where_iGeaendert}{condition for field \emph{iGeaendert}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_iID}{\code{select} field \emph{iID}?}

\item{select_iEffektID}{\code{select} field \emph{iEffektID}?}

\item{select_iName}{\code{select} field \emph{iName}?}

\item{select_iEinheit}{\code{select} field \emph{iEinheit}?}

\item{select_iBeschreibung}{\code{select} field \emph{iBeschreibung}?}

\item{select_iAngelegtAm}{\code{select} field \emph{iAngelegtAm}?}

\item{select_iAngelegtVonPersonID}{\code{select} field \emph{iAngelegtVonPersonID}?}

\item{select_iGeaendert}{\code{select} field \emph{iGeaendert}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_iID}{order by field \emph{iID}?}

\item{orderBy_iEffektID}{order by field \emph{iEffektID}?}

\item{orderBy_iName}{order by field \emph{iName}?}

\item{orderBy_iEinheit}{order by field \emph{iEinheit}?}

\item{orderBy_iBeschreibung}{order by field \emph{iBeschreibung}?}

\item{orderBy_iAngelegtAm}{order by field \emph{iAngelegtAm}?}

\item{orderBy_iAngelegtVonPersonID}{order by field \emph{iAngelegtVonPersonID}?}

\item{orderBy_iGeaendert}{order by field \emph{iGeaendert}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblindikator
}
