% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_indikatorwert}
\alias{kuras_indikatorwert}
\title{select records from tblindikatorwert}
\usage{
kuras_indikatorwert(
  where = NULL,
  where_iwID = NA,
  where_iwAnlageID = NA,
  where_iwIndikatorID = NA,
  where_iwReferenzGruppeID = NA,
  where_iwMittelwert = NA,
  where_iwMinWert = NA,
  where_iwMaxWert = NA,
  where_iwMedian = NA,
  where_iwStandardabweichung = NA,
  where_iwAnzahlWerte = NA,
  where_iwZeitbezugBeginn = NA,
  where_iwZeitbezugEnde = NA,
  where_iwErhebungsmethode = NA,
  select = NULL,
  select_iwID = NA,
  select_iwAnlageID = NA,
  select_iwIndikatorID = NA,
  select_iwReferenzGruppeID = NA,
  select_iwMittelwert = NA,
  select_iwMinWert = NA,
  select_iwMaxWert = NA,
  select_iwMedian = NA,
  select_iwStandardabweichung = NA,
  select_iwAnzahlWerte = NA,
  select_iwZeitbezugBeginn = NA,
  select_iwZeitbezugEnde = NA,
  select_iwErhebungsmethode = NA,
  orderBy = NULL,
  orderBy_iwID = NA,
  orderBy_iwAnlageID = NA,
  orderBy_iwIndikatorID = NA,
  orderBy_iwReferenzGruppeID = NA,
  orderBy_iwMittelwert = NA,
  orderBy_iwMinWert = NA,
  orderBy_iwMaxWert = NA,
  orderBy_iwMedian = NA,
  orderBy_iwStandardabweichung = NA,
  orderBy_iwAnzahlWerte = NA,
  orderBy_iwZeitbezugBeginn = NA,
  orderBy_iwZeitbezugEnde = NA,
  orderBy_iwErhebungsmethode = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_iwID}{condition for field \emph{iwID}: Enter a field description here}

\item{where_iwAnlageID}{condition for field \emph{iwAnlageID}: Enter a field description here}

\item{where_iwIndikatorID}{condition for field \emph{iwIndikatorID}: Enter a field description here}

\item{where_iwReferenzGruppeID}{condition for field \emph{iwReferenzGruppeID}: Enter a field description here}

\item{where_iwMittelwert}{condition for field \emph{iwMittelwert}: Enter a field description here}

\item{where_iwMinWert}{condition for field \emph{iwMinWert}: Enter a field description here}

\item{where_iwMaxWert}{condition for field \emph{iwMaxWert}: Enter a field description here}

\item{where_iwMedian}{condition for field \emph{iwMedian}: Enter a field description here}

\item{where_iwStandardabweichung}{condition for field \emph{iwStandardabweichung}: Enter a field description here}

\item{where_iwAnzahlWerte}{condition for field \emph{iwAnzahlWerte}: Enter a field description here}

\item{where_iwZeitbezugBeginn}{condition for field \emph{iwZeitbezugBeginn}: Enter a field description here}

\item{where_iwZeitbezugEnde}{condition for field \emph{iwZeitbezugEnde}: Enter a field description here}

\item{where_iwErhebungsmethode}{condition for field \emph{iwErhebungsmethode}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_iwID}{\code{select} field \emph{iwID}?}

\item{select_iwAnlageID}{\code{select} field \emph{iwAnlageID}?}

\item{select_iwIndikatorID}{\code{select} field \emph{iwIndikatorID}?}

\item{select_iwReferenzGruppeID}{\code{select} field \emph{iwReferenzGruppeID}?}

\item{select_iwMittelwert}{\code{select} field \emph{iwMittelwert}?}

\item{select_iwMinWert}{\code{select} field \emph{iwMinWert}?}

\item{select_iwMaxWert}{\code{select} field \emph{iwMaxWert}?}

\item{select_iwMedian}{\code{select} field \emph{iwMedian}?}

\item{select_iwStandardabweichung}{\code{select} field \emph{iwStandardabweichung}?}

\item{select_iwAnzahlWerte}{\code{select} field \emph{iwAnzahlWerte}?}

\item{select_iwZeitbezugBeginn}{\code{select} field \emph{iwZeitbezugBeginn}?}

\item{select_iwZeitbezugEnde}{\code{select} field \emph{iwZeitbezugEnde}?}

\item{select_iwErhebungsmethode}{\code{select} field \emph{iwErhebungsmethode}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_iwID}{order by field \emph{iwID}?}

\item{orderBy_iwAnlageID}{order by field \emph{iwAnlageID}?}

\item{orderBy_iwIndikatorID}{order by field \emph{iwIndikatorID}?}

\item{orderBy_iwReferenzGruppeID}{order by field \emph{iwReferenzGruppeID}?}

\item{orderBy_iwMittelwert}{order by field \emph{iwMittelwert}?}

\item{orderBy_iwMinWert}{order by field \emph{iwMinWert}?}

\item{orderBy_iwMaxWert}{order by field \emph{iwMaxWert}?}

\item{orderBy_iwMedian}{order by field \emph{iwMedian}?}

\item{orderBy_iwStandardabweichung}{order by field \emph{iwStandardabweichung}?}

\item{orderBy_iwAnzahlWerte}{order by field \emph{iwAnzahlWerte}?}

\item{orderBy_iwZeitbezugBeginn}{order by field \emph{iwZeitbezugBeginn}?}

\item{orderBy_iwZeitbezugEnde}{order by field \emph{iwZeitbezugEnde}?}

\item{orderBy_iwErhebungsmethode}{order by field \emph{iwErhebungsmethode}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblindikatorwert
}
