% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_institution}
\alias{kuras_institution}
\title{select records from tblinstitution}
\usage{
kuras_institution(
  where = NULL,
  where_inID = NA,
  where_inName = NA,
  where_inKurzname = NA,
  where_inAdresse = NA,
  where_inBeschreibung = NA,
  select = NULL,
  select_inID = NA,
  select_inName = NA,
  select_inKurzname = NA,
  select_inAdresse = NA,
  select_inBeschreibung = NA,
  orderBy = NULL,
  orderBy_inID = NA,
  orderBy_inName = NA,
  orderBy_inKurzname = NA,
  orderBy_inAdresse = NA,
  orderBy_inBeschreibung = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_inID}{condition for field \emph{inID}: Enter a field description here}

\item{where_inName}{condition for field \emph{inName}: Enter a field description here}

\item{where_inKurzname}{condition for field \emph{inKurzname}: Enter a field description here}

\item{where_inAdresse}{condition for field \emph{inAdresse}: Enter a field description here}

\item{where_inBeschreibung}{condition for field \emph{inBeschreibung}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_inID}{\code{select} field \emph{inID}?}

\item{select_inName}{\code{select} field \emph{inName}?}

\item{select_inKurzname}{\code{select} field \emph{inKurzname}?}

\item{select_inAdresse}{\code{select} field \emph{inAdresse}?}

\item{select_inBeschreibung}{\code{select} field \emph{inBeschreibung}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_inID}{order by field \emph{inID}?}

\item{orderBy_inName}{order by field \emph{inName}?}

\item{orderBy_inKurzname}{order by field \emph{inKurzname}?}

\item{orderBy_inAdresse}{order by field \emph{inAdresse}?}

\item{orderBy_inBeschreibung}{order by field \emph{inBeschreibung}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblinstitution
}
