% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_kommentar}
\alias{kuras_kommentar}
\title{select records from tblkommentar}
\usage{
kuras_kommentar(
  where = NULL,
  where_kID = NA,
  where_kPersonID = NA,
  where_kDatum = NA,
  where_kBezugstabelle = NA,
  where_kDatensatzID = NA,
  where_kKommentar = NA,
  select = NULL,
  select_kID = NA,
  select_kPersonID = NA,
  select_kDatum = NA,
  select_kBezugstabelle = NA,
  select_kDatensatzID = NA,
  select_kKommentar = NA,
  orderBy = NULL,
  orderBy_kID = NA,
  orderBy_kPersonID = NA,
  orderBy_kDatum = NA,
  orderBy_kBezugstabelle = NA,
  orderBy_kDatensatzID = NA,
  orderBy_kKommentar = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_kID}{condition for field \emph{kID}: Enter a field description here}

\item{where_kPersonID}{condition for field \emph{kPersonID}: Enter a field description here}

\item{where_kDatum}{condition for field \emph{kDatum}: Enter a field description here}

\item{where_kBezugstabelle}{condition for field \emph{kBezugstabelle}: Enter a field description here}

\item{where_kDatensatzID}{condition for field \emph{kDatensatzID}: Enter a field description here}

\item{where_kKommentar}{condition for field \emph{kKommentar}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_kID}{\code{select} field \emph{kID}?}

\item{select_kPersonID}{\code{select} field \emph{kPersonID}?}

\item{select_kDatum}{\code{select} field \emph{kDatum}?}

\item{select_kBezugstabelle}{\code{select} field \emph{kBezugstabelle}?}

\item{select_kDatensatzID}{\code{select} field \emph{kDatensatzID}?}

\item{select_kKommentar}{\code{select} field \emph{kKommentar}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_kID}{order by field \emph{kID}?}

\item{orderBy_kPersonID}{order by field \emph{kPersonID}?}

\item{orderBy_kDatum}{order by field \emph{kDatum}?}

\item{orderBy_kBezugstabelle}{order by field \emph{kBezugstabelle}?}

\item{orderBy_kDatensatzID}{order by field \emph{kDatensatzID}?}

\item{orderBy_kKommentar}{order by field \emph{kKommentar}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblkommentar
}
