% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_koordinatensystem}
\alias{kuras_koordinatensystem}
\title{select records from tblkoordinatensystem}
\usage{
kuras_koordinatensystem(
  where = NULL,
  where_koID = NA,
  where_koName = NA,
  select = NULL,
  select_koID = NA,
  select_koName = NA,
  orderBy = NULL,
  orderBy_koID = NA,
  orderBy_koName = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_koID}{condition for field \emph{koID}: Enter a field description here}

\item{where_koName}{condition for field \emph{koName}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_koID}{\code{select} field \emph{koID}?}

\item{select_koName}{\code{select} field \emph{koName}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_koID}{order by field \emph{koID}?}

\item{orderBy_koName}{order by field \emph{koName}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblkoordinatensystem
}
