% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_massnahmeebene}
\alias{kuras_massnahmeebene}
\title{select records from tblmassnahmeebene}
\usage{
kuras_massnahmeebene(
  where = NULL,
  where_meID = NA,
  where_meName = NA,
  where_meBeschreibung = NA,
  select = NULL,
  select_meID = NA,
  select_meName = NA,
  select_meBeschreibung = NA,
  orderBy = NULL,
  orderBy_meID = NA,
  orderBy_meName = NA,
  orderBy_meBeschreibung = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_meID}{condition for field \emph{meID}: Enter a field description here}

\item{where_meName}{condition for field \emph{meName}: Enter a field description here}

\item{where_meBeschreibung}{condition for field \emph{meBeschreibung}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_meID}{\code{select} field \emph{meID}?}

\item{select_meName}{\code{select} field \emph{meName}?}

\item{select_meBeschreibung}{\code{select} field \emph{meBeschreibung}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_meID}{order by field \emph{meID}?}

\item{orderBy_meName}{order by field \emph{meName}?}

\item{orderBy_meBeschreibung}{order by field \emph{meBeschreibung}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblmassnahmeebene
}
