% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_massnahmenkategorie}
\alias{kuras_massnahmenkategorie}
\title{select records from tblmassnahmenkategorie}
\usage{
kuras_massnahmenkategorie(
  where = NULL,
  where_mkID = NA,
  where_mkName = NA,
  select = NULL,
  select_mkID = NA,
  select_mkName = NA,
  orderBy = NULL,
  orderBy_mkID = NA,
  orderBy_mkName = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_mkID}{condition for field \emph{mkID}: Enter a field description here}

\item{where_mkName}{condition for field \emph{mkName}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_mkID}{\code{select} field \emph{mkID}?}

\item{select_mkName}{\code{select} field \emph{mkName}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_mkID}{order by field \emph{mkID}?}

\item{orderBy_mkName}{order by field \emph{mkName}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblmassnahmenkategorie
}
