% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_massnahmetyp}
\alias{kuras_massnahmetyp}
\title{select records from tblmassnahmetyp}
\usage{
kuras_massnahmetyp(
  where = NULL,
  where_mID = NA,
  where_mName = NA,
  where_mBeschreibung = NA,
  where_mMassnahmenKategorieID = NA,
  select = NULL,
  select_mID = NA,
  select_mName = NA,
  select_mBeschreibung = NA,
  select_mMassnahmenKategorieID = NA,
  orderBy = NULL,
  orderBy_mID = NA,
  orderBy_mName = NA,
  orderBy_mBeschreibung = NA,
  orderBy_mMassnahmenKategorieID = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_mID}{condition for field \emph{mID}: Enter a field description here}

\item{where_mName}{condition for field \emph{mName}: Enter a field description here}

\item{where_mBeschreibung}{condition for field \emph{mBeschreibung}: Enter a field description here}

\item{where_mMassnahmenKategorieID}{condition for field \emph{mMassnahmenKategorieID}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_mID}{\code{select} field \emph{mID}?}

\item{select_mName}{\code{select} field \emph{mName}?}

\item{select_mBeschreibung}{\code{select} field \emph{mBeschreibung}?}

\item{select_mMassnahmenKategorieID}{\code{select} field \emph{mMassnahmenKategorieID}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_mID}{order by field \emph{mID}?}

\item{orderBy_mName}{order by field \emph{mName}?}

\item{orderBy_mBeschreibung}{order by field \emph{mBeschreibung}?}

\item{orderBy_mMassnahmenKategorieID}{order by field \emph{mMassnahmenKategorieID}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblmassnahmetyp
}
