% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_person}
\alias{kuras_person}
\title{select records from tblperson}
\usage{
kuras_person(
  where = NULL,
  where_pnID = NA,
  where_pnInitialien = NA,
  where_pnVorname = NA,
  where_pnNachname = NA,
  where_pnInstitutionID = NA,
  where_pnAdresse = NA,
  where_pnTelefon = NA,
  select = NULL,
  select_pnID = NA,
  select_pnInitialien = NA,
  select_pnVorname = NA,
  select_pnNachname = NA,
  select_pnInstitutionID = NA,
  select_pnAdresse = NA,
  select_pnTelefon = NA,
  orderBy = NULL,
  orderBy_pnID = NA,
  orderBy_pnInitialien = NA,
  orderBy_pnVorname = NA,
  orderBy_pnNachname = NA,
  orderBy_pnInstitutionID = NA,
  orderBy_pnAdresse = NA,
  orderBy_pnTelefon = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_pnID}{condition for field \emph{pnID}: Enter a field description here}

\item{where_pnInitialien}{condition for field \emph{pnInitialien}: Enter a field description here}

\item{where_pnVorname}{condition for field \emph{pnVorname}: Enter a field description here}

\item{where_pnNachname}{condition for field \emph{pnNachname}: Enter a field description here}

\item{where_pnInstitutionID}{condition for field \emph{pnInstitutionID}: Enter a field description here}

\item{where_pnAdresse}{condition for field \emph{pnAdresse}: Enter a field description here}

\item{where_pnTelefon}{condition for field \emph{pnTelefon}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_pnID}{\code{select} field \emph{pnID}?}

\item{select_pnInitialien}{\code{select} field \emph{pnInitialien}?}

\item{select_pnVorname}{\code{select} field \emph{pnVorname}?}

\item{select_pnNachname}{\code{select} field \emph{pnNachname}?}

\item{select_pnInstitutionID}{\code{select} field \emph{pnInstitutionID}?}

\item{select_pnAdresse}{\code{select} field \emph{pnAdresse}?}

\item{select_pnTelefon}{\code{select} field \emph{pnTelefon}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_pnID}{order by field \emph{pnID}?}

\item{orderBy_pnInitialien}{order by field \emph{pnInitialien}?}

\item{orderBy_pnVorname}{order by field \emph{pnVorname}?}

\item{orderBy_pnNachname}{order by field \emph{pnNachname}?}

\item{orderBy_pnInstitutionID}{order by field \emph{pnInstitutionID}?}

\item{orderBy_pnAdresse}{order by field \emph{pnAdresse}?}

\item{orderBy_pnTelefon}{order by field \emph{pnTelefon}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblperson
}
