% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_projekt}
\alias{kuras_projekt}
\title{select records from tblprojekt}
\usage{
kuras_projekt(
  where = NULL,
  where_pID = NA,
  where_pName = NA,
  where_pBeginn = NA,
  where_pEnde = NA,
  where_pDurchfuehrungInstitutionID = NA,
  where_pAuftraggeberInstitutionID = NA,
  where_pBeschreibung = NA,
  where_pProjektblattnummer = NA,
  select = NULL,
  select_pID = NA,
  select_pName = NA,
  select_pBeginn = NA,
  select_pEnde = NA,
  select_pDurchfuehrungInstitutionID = NA,
  select_pAuftraggeberInstitutionID = NA,
  select_pBeschreibung = NA,
  select_pProjektblattnummer = NA,
  orderBy = NULL,
  orderBy_pID = NA,
  orderBy_pName = NA,
  orderBy_pBeginn = NA,
  orderBy_pEnde = NA,
  orderBy_pDurchfuehrungInstitutionID = NA,
  orderBy_pAuftraggeberInstitutionID = NA,
  orderBy_pBeschreibung = NA,
  orderBy_pProjektblattnummer = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_pID}{condition for field \emph{pID}: Enter a field description here}

\item{where_pName}{condition for field \emph{pName}: Enter a field description here}

\item{where_pBeginn}{condition for field \emph{pBeginn}: Enter a field description here}

\item{where_pEnde}{condition for field \emph{pEnde}: Enter a field description here}

\item{where_pDurchfuehrungInstitutionID}{condition for field \emph{pDurchfuehrungInstitutionID}: Enter a field description here}

\item{where_pAuftraggeberInstitutionID}{condition for field \emph{pAuftraggeberInstitutionID}: Enter a field description here}

\item{where_pBeschreibung}{condition for field \emph{pBeschreibung}: Enter a field description here}

\item{where_pProjektblattnummer}{condition for field \emph{pProjektblattnummer}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_pID}{\code{select} field \emph{pID}?}

\item{select_pName}{\code{select} field \emph{pName}?}

\item{select_pBeginn}{\code{select} field \emph{pBeginn}?}

\item{select_pEnde}{\code{select} field \emph{pEnde}?}

\item{select_pDurchfuehrungInstitutionID}{\code{select} field \emph{pDurchfuehrungInstitutionID}?}

\item{select_pAuftraggeberInstitutionID}{\code{select} field \emph{pAuftraggeberInstitutionID}?}

\item{select_pBeschreibung}{\code{select} field \emph{pBeschreibung}?}

\item{select_pProjektblattnummer}{\code{select} field \emph{pProjektblattnummer}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_pID}{order by field \emph{pID}?}

\item{orderBy_pName}{order by field \emph{pName}?}

\item{orderBy_pBeginn}{order by field \emph{pBeginn}?}

\item{orderBy_pEnde}{order by field \emph{pEnde}?}

\item{orderBy_pDurchfuehrungInstitutionID}{order by field \emph{pDurchfuehrungInstitutionID}?}

\item{orderBy_pAuftraggeberInstitutionID}{order by field \emph{pAuftraggeberInstitutionID}?}

\item{orderBy_pBeschreibung}{order by field \emph{pBeschreibung}?}

\item{orderBy_pProjektblattnummer}{order by field \emph{pProjektblattnummer}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblprojekt
}
