% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_referenz}
\alias{kuras_referenz}
\title{select records from tblreferenz}
\usage{
kuras_referenz(
  where = NULL,
  where_rID = NA,
  where_rNachname = NA,
  where_rVorname = NA,
  where_rJahr = NA,
  where_rArt = NA,
  where_rTitel = NA,
  where_rTitelMedium = NA,
  where_rZusatz = NA,
  where_rBeschreibung = NA,
  where_rQualitaet = NA,
  select = NULL,
  select_rID = NA,
  select_rNachname = NA,
  select_rVorname = NA,
  select_rJahr = NA,
  select_rArt = NA,
  select_rTitel = NA,
  select_rTitelMedium = NA,
  select_rZusatz = NA,
  select_rBeschreibung = NA,
  select_rQualitaet = NA,
  orderBy = NULL,
  orderBy_rID = NA,
  orderBy_rNachname = NA,
  orderBy_rVorname = NA,
  orderBy_rJahr = NA,
  orderBy_rArt = NA,
  orderBy_rTitel = NA,
  orderBy_rTitelMedium = NA,
  orderBy_rZusatz = NA,
  orderBy_rBeschreibung = NA,
  orderBy_rQualitaet = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_rID}{condition for field \emph{rID}: Enter a field description here}

\item{where_rNachname}{condition for field \emph{rNachname}: Enter a field description here}

\item{where_rVorname}{condition for field \emph{rVorname}: Enter a field description here}

\item{where_rJahr}{condition for field \emph{rJahr}: Enter a field description here}

\item{where_rArt}{condition for field \emph{rArt}: Enter a field description here}

\item{where_rTitel}{condition for field \emph{rTitel}: Enter a field description here}

\item{where_rTitelMedium}{condition for field \emph{rTitelMedium}: Enter a field description here}

\item{where_rZusatz}{condition for field \emph{rZusatz}: Enter a field description here}

\item{where_rBeschreibung}{condition for field \emph{rBeschreibung}: Enter a field description here}

\item{where_rQualitaet}{condition for field \emph{rQualitaet}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_rID}{\code{select} field \emph{rID}?}

\item{select_rNachname}{\code{select} field \emph{rNachname}?}

\item{select_rVorname}{\code{select} field \emph{rVorname}?}

\item{select_rJahr}{\code{select} field \emph{rJahr}?}

\item{select_rArt}{\code{select} field \emph{rArt}?}

\item{select_rTitel}{\code{select} field \emph{rTitel}?}

\item{select_rTitelMedium}{\code{select} field \emph{rTitelMedium}?}

\item{select_rZusatz}{\code{select} field \emph{rZusatz}?}

\item{select_rBeschreibung}{\code{select} field \emph{rBeschreibung}?}

\item{select_rQualitaet}{\code{select} field \emph{rQualitaet}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_rID}{order by field \emph{rID}?}

\item{orderBy_rNachname}{order by field \emph{rNachname}?}

\item{orderBy_rVorname}{order by field \emph{rVorname}?}

\item{orderBy_rJahr}{order by field \emph{rJahr}?}

\item{orderBy_rArt}{order by field \emph{rArt}?}

\item{orderBy_rTitel}{order by field \emph{rTitel}?}

\item{orderBy_rTitelMedium}{order by field \emph{rTitelMedium}?}

\item{orderBy_rZusatz}{order by field \emph{rZusatz}?}

\item{orderBy_rBeschreibung}{order by field \emph{rBeschreibung}?}

\item{orderBy_rQualitaet}{order by field \emph{rQualitaet}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblreferenz
}
