% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_referenzgruppe}
\alias{kuras_referenzgruppe}
\title{select records from tblreferenzgruppe}
\usage{
kuras_referenzgruppe(
  where = NULL,
  where_rgID = NA,
  select = NULL,
  select_rgID = NA,
  orderBy = NULL,
  orderBy_rgID = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_rgID}{condition for field \emph{rgID}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_rgID}{\code{select} field \emph{rgID}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_rgID}{order by field \emph{rgID}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblreferenzgruppe
}
