% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_referenzgruppereferenz}
\alias{kuras_referenzgruppereferenz}
\title{select records from tblreferenzgruppereferenz}
\usage{
kuras_referenzgruppereferenz(
  where = NULL,
  where_rrID = NA,
  where_rrReferenzGruppeID = NA,
  where_rrReferenzID = NA,
  select = NULL,
  select_rrID = NA,
  select_rrReferenzGruppeID = NA,
  select_rrReferenzID = NA,
  orderBy = NULL,
  orderBy_rrID = NA,
  orderBy_rrReferenzGruppeID = NA,
  orderBy_rrReferenzID = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_rrID}{condition for field \emph{rrID}: Enter a field description here}

\item{where_rrReferenzGruppeID}{condition for field \emph{rrReferenzGruppeID}: Enter a field description here}

\item{where_rrReferenzID}{condition for field \emph{rrReferenzID}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_rrID}{\code{select} field \emph{rrID}?}

\item{select_rrReferenzGruppeID}{\code{select} field \emph{rrReferenzGruppeID}?}

\item{select_rrReferenzID}{\code{select} field \emph{rrReferenzID}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_rrID}{order by field \emph{rrID}?}

\item{orderBy_rrReferenzGruppeID}{order by field \emph{rrReferenzGruppeID}?}

\item{orderBy_rrReferenzID}{order by field \emph{rrReferenzID}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblreferenzgruppereferenz
}
