% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasDbInterface.R
\name{kuras_standort}
\alias{kuras_standort}
\title{select records from tblstandort}
\usage{
kuras_standort(
  where = NULL,
  where_oID = NA,
  where_oName = NA,
  where_oOrt = NA,
  select = NULL,
  select_oID = NA,
  select_oName = NA,
  select_oOrt = NA,
  orderBy = NULL,
  orderBy_oID = NA,
  orderBy_oName = NA,
  orderBy_oOrt = NA,
  db = "KURAS",
  ...
)
}
\arguments{
\item{where}{WHERE condition}

\item{where_oID}{condition for field \emph{oID}: Enter a field description here}

\item{where_oName}{condition for field \emph{oName}: Enter a field description here}

\item{where_oOrt}{condition for field \emph{oOrt}: Enter a field description here}

\item{select}{fields/expressions to \code{select}}

\item{select_oID}{\code{select} field \emph{oID}?}

\item{select_oName}{\code{select} field \emph{oName}?}

\item{select_oOrt}{\code{select} field \emph{oOrt}?}

\item{orderBy}{ORDER BY clause of SQL expression}

\item{orderBy_oID}{order by field \emph{oID}?}

\item{orderBy_oName}{order by field \emph{oName}?}

\item{orderBy_oOrt}{order by field \emph{oOrt}?}

\item{db}{name of ODBC data source, default: \code{"KURAS"}}

\item{\dots}{additional arguments passed to \code{kwb.db::selectFromTable}}
}
\description{
\code{select} and filter records from tblstandort
}
