% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasReferencesToRIS.R
\name{kurasReferencesToRIS}
\alias{kurasReferencesToRIS}
\title{Write References to file in RIS (Research Information Systems) Format}
\usage{
kurasReferencesToRIS(references)
}
\arguments{
\item{references}{data frame with columns \code{rNachname}, \code{rVorname}, 
\code{rTitel}, \code{Medium}, \code{rZusatz}, \code{rJahr}, 
\code{rBeschreibung}}
}
\value{
path to the file created in the tempdir() folder
}
\description{
Write References to file in RIS (Research Information Systems) Format
}
\examples{
\dontrun{
# Get all entries from KURAS table "tblReferenz"
references <- kuras_referenz()
  
# Generate RIS-file in temporary folder
file <- kurasReferencesToRIS(references)

# Open the file in the Windows Explorer
# kwb.utils::hsOpenWindowsExplorer(file)
}

}
